feat: search both origin and canonicalized directory of `current_exe`
In recent changes, we force `current_exe` to be canonicalized.
However, for homebrew, `maa` is installed in separate directory,
and linked to `/usr/local/bin/maa` or `/opt/homebrew/bin/maa`.
To search library and resource also installed by homebrew,
we need to search the origin directory of `current_exe`.
So we search both origin and canonicalized directory of `current_exe`.