ci: use tag pattern to filter out pre-releases
So far, we have been using the `compare_base` output from the
`parse_version.sh` script to determine the range of commits to be
included in the changelog. This not works well,
because the pre-release also splits the commits into multiple groups.
For example, if the last stable release is `v0.1.0`,
and last beta release is `v0.1.1-beta.1`, then the changelog from head
will include all commits from `v0.1.0` to head, but
they will be split into two groups: `v0.1.0..v0.1.1-beta.1` and
`v0.1.1-beta.1..HEAD`. This is not what we want.

So instead of using `compare_base`, we use the `tag_pattern` to filter
out pre-releases. This will ensure that the changelog compares to the
last stable release, and they are grouped into a single group.