test: fix test failure on CI caused by create user resource dir
The CI fails with message "Cannot create a file when that file already exists",
which might be caused by the test trying to create the user resource directory
twice. However, I have no idea why this happens on CI, because lazy_static is
supposed to be initialized only once.
This commit tries to fix this issue by using a static atomic bool to lock the
creation of user resource directory.