feat: improve optional value (#280)
- Rename `OptionalInput` to `Optional`;
- Rename `deps` to `conditions` in `Optional`;
- Rename `input` to `value` in `Optional` and can be any boxed
`MAAValue`;
- Fix initialize order of values in an object:
The current sort implementation is incorrect. The `sort_by` function is
only suitable for sorting elements with a total order, whereas the
dependencies exhibit a partial order. In the current implementation, we
define a total order that combines dependencies and key names, which is
not transitive, even in the absence of circular dependencies. For
instance, if A depends on B, B depends on C, and C depends on D, then it
should be A < B < C < D. However, C and A are both `Optional` and do not
directly depend on each other. Consequently, they will be compared by
their keys, and the order may be C > A, which contradicts the
dependencies. We should not use keys to compare elements, but rather the
dependencies. Although the dependencies are a partial order, we can sort
them using a topological sort (depth-first search).
- Improve the macro `object!` that can be used to create optional.