fix: use self-replace to update maa-cli self
On windows, the maa-cli binary is locked by the system when it is running.
This makes it impossible to update the binary directly.
To solve this problem, we can use the self-replace crate to update the binary.

Besides, constcat crate is added to create some platform-specific file names.
The extract function is slightly modified to accept a closure that accepts a
Cow<Path> instead of a &Path.